Set Hook switches to assumed_state=true
Hook can't poll for status, so the switches should be treated as assumed
state. People can always customize it back to false if they like the
toggle better.

Also adds a workaround for a crash from aiohttp caused by the Hook
endpoint returning json but with a content-type: text/html header.